# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memora API (mock)
  version: "1.0.0"
  description: |
    API mock do Memora — Express 5, estado em memória (reinicia com o processo, sem banco de
    dados). Reproduz as regras de negócio do Escopo Consolidado v2 para servir de backend ao
    protótipo navegável; não é a API de produção.

    Todas as rotas de negócio vivem sob `/v1`. A autenticação é por sessão simples (token opaco
    em `Authorization: Bearer <token>`), sem JWT nem refresh — ver
    [ADR-0008](./adr/0008-autenticacao-por-sessao-simples.md).

    Cada área do produto (aluno, responsável, B2B, back-office, parceiro) faz login com sua
    própria rota `POST .../auth/login` e recebe um token cujo `role` restringe o que pode
    acessar (`requireRole` em `backend/src/lib/auth.js`).
  contact:
    name: Memora — engenharia
  license:
    name: Uso interno
servers:
  - url: http://localhost:3180/v1
    description: Desenvolvimento local (porta configurável via `PORT`)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token de sessão retornado por qualquer rota `auth/login`. Sessões expiram em 30 dias
        sem uso (deslizante); usar em `me`, `parental`, `b2b`, `admin` e `partner` é mutuamente
        exclusivo por `role`.
  parameters:
    deckId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: ID do deck.
...
  responses:
    Unauthorized:
      description: Sessão ausente, inválida ou expirada.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: unauthorized, message: "Sessão inválida ou expirada." }
...
    NotFound:
      description: Recurso não encontrado (ou não pertence ao usuário autenticado).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validation:
      description: Corpo da requisição inválido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "Slug estável da causa, ex.: validation, not_found, guardian_required." }
        message: { type: string, description: "Mensagem em PT-BR, pronta para exibição." }
      additionalProperties: true
      description: Formato uniforme de erro em toda a API (ver `fail()` em `backend/src/lib/helpers.js`).
    User:
      type: object
      description: Perfil público do aluno logado (`publicUser()` em `backend/src/lib/helpers.js`).
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        birthDate: { type: string, format: date }
        isMinor: { type: boolean }
        guardianId: { type: string, nullable: true }
        plan: { type: string, enum: [free, premium] }
        locale: { type: string, example: pt-BR }
        timezone: { type: string, example: America/Sao_Paulo }
        verified:
          type: object
          properties: { email: { type: boolean }, phone: { type: boolean } }
        canEarn: { type: boolean, description: "true somente com e-mail e telefone confirmados." }
        ageVerification:
          type: object
          properties:
            status: { type: string, enum: [pendente, aguardando_responsavel, aprovado] }
            method: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        deactivatedAt: { type: string, format: date-time, nullable: true }
        publicProfile: { type: object, additionalProperties: true }
        settings: { type: object, additionalProperties: true }
        guardianPermissions: { type: object, nullable: true, additionalProperties: true }
        streak: { type: integer }
        studyDays: { type: integer }
        balance: { type: integer, description: "Saldo de Neurônios." }
        onboardingDone: { type: boolean }
...
    LedgerEntry:
      type: object
      description: Lançamento imutável do ledger de Neurônios (`backend/src/lib/ledger.js`).
      properties:
        id: { type: string }
        ts: { type: string, format: date-time }
        userId: { type: string }
        amount: { type: integer, description: "Positivo = crédito, negativo = débito." }
        type:
          type: string
          enum: [daily_access, goal_bonus, objective_outcome, clone_bonus, referral_bonus, coupon_redeem, coupon_expired_refund, proto_adjustment]
        meta: { type: object, additionalProperties: true }
        balanceAfter: { type: integer }
...
paths:
...
  /me:
    get:
      tags: [Perfil]
      summary: Dados da própria conta
      description: "Credita +1 Neurônio de acesso diário (uma vez por dia, se verificado e não desativado)."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  dailyCredit: { $ref: '#/components/schemas/LedgerEntry' }
                  notifications: { type: object, additionalProperties: true }
                  consents: { type: object, additionalProperties: true }
                  subscription: { type: object, additionalProperties: true }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /me/settings:
    patch:
      tags: [Perfil]
      summary: Atualiza nome, idioma, fuso, meta diária, intensidade de notificação e acessibilidade
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                locale: { type: string }
                timezone: { type: string }
                intensity: { type: string, enum: [leve, padrao, intenso], description: "Contas de menor ficam travadas em leve." }
                reminderTime: { type: string, example: "20:00" }
                push: { type: boolean }
                dailyGoal: { type: integer, minimum: 5, maximum: 200 }
                fatigueDetection: { type: boolean }
                accessibility: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
        "403": { description: "`intensity_locked` para contas de menor.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /me/sessions/{id}:
    delete:
      tags: [Perfil]
      summary: Encerra uma sessão específica
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "409": { description: "Não é possível encerrar a sessão atual por aqui." }
  /me/password:
    post:
      tags: [Perfil]
      summary: Troca de senha (encerra as outras sessões)
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [current, next], properties: { current: { type: string }, next: { type: string, minLength: 8 } } }
      responses:
        "200": { description: OK }
        "422": { $ref: '#/components/responses/Validation' }
  /me/consents:
    get:
      tags: [Perfil]
      summary: Consentimentos (escola, WhatsApp por categoria)
      responses: { "200": { description: OK } }
    patch:
      tags: [Perfil]
      summary: Atualiza consentimentos
      description: "Menores sem liberação do responsável recebem `guardian_required` (403) em vez de aplicar."
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schoolShare: { type: boolean }
                whatsapp: { type: object, additionalProperties: { type: boolean } }
      responses:
        "200": { description: OK }
        "403": { description: guardian_required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /me/guardian-requests:
    get:
      tags: [Perfil]
      summary: Lista solicitações pendentes ao responsável (só para menores)
      responses: { "200": { description: OK } }
    post:
      tags: [Perfil]
      summary: Cria uma solicitação de liberação ao responsável
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [community, publicLink, externalMembers, schoolShare, whatsappInfo] }
                detail: { type: string }
      responses:
        "201": { description: Criada }
        "422": { $ref: '#/components/responses/Validation' }

  # ----------------------------------------------------------------- Estudo
...
  /decks/{id}:
    get:
      tags: [Decks]
      summary: Detalhe do deck, cards e menu de ações disponíveis
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Decks]
      summary: Atualiza flags (difícil/foco/complementar/fixado), nome, categoria, tags ou pasta
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { description: "`pin_limit` — limite de decks fixados do plano atingido." }
    delete:
      tags: [Decks]
      summary: Exclui o deck (soft delete)
      parameters: [{ $ref: '#/components/parameters/deckId' }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /noa/suggestions/{id}/respond:
    post:
      tags: [Nôa]
      summary: Aceita ou dispensa uma sugestão da Nôa (até 5 sugestões/dia; 1 ação autônoma/dia)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { accepted: { type: boolean }, feedback: { type: string } } } } } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /noa/decisions/{id}/revert:
    post:
      tags: [Nôa]
      summary: Reverte uma decisão autônoma reversível da Nôa
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { description: not_reversible }

  # ------------------------------------------------------------------ Público
...